requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
baron==0.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.7
django-allauth==0.45.0
django-cache-url==3.2.3
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environs==0.1.1
django-heroku==0.3.1
django-jet==1.0.8
django-phonenumber-field==5.2.0
environs==9.3.4
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
marshmallow==3.13.0
oauthlib==3.1.1
packaging==21.0
phonenumbers==8.12.34
Pillow==8.3.2
pluggy==1.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
redbaron==0.9.2
requests==2.26.0
requests-oauthlib==1.3.0
rply==0.7.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
